requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Deep learning
torch>=2.0.0
dhg>=0.9.3  # DeepHypergraph library

# Machine learning
scikit-learn>=1.0.0
lightgbm>=3.3.0

# Chemistry and bioinformatics
rdkit>=2022.03.1  # For molecular fingerprints and Tanimoto similarity
biopython>=1.79  # For BLAST and sequence analysis

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
networkx>=2.6.0

# Utilities
tqdm>=4.62.0  # Progress bars
pyyaml>=5.4.0  # Configuration file parsing
joblib>=1.1.0  # Parallel processing

# Development (optional)
pytest>=6.2.0
black>=21.9b0
flake8>=3.9.0
